+++ lib/passport_strategy.js
@@ -3,10 +3,11 @@
 // TODO: SET THE HTTP PATH ON CALLBACKURL as ENV VARIABLE
 
 var FacebookStrategy = require('passport-facebook').Strategy;
-var User             = require('../models/User.js');
+var User = require('../models/User.js');
 var fs = require('fs');
 var request = require('superagent');
 var Card = require('../models/Card.js');
+var getFacebookFriends = require('getFacebookFriends');
 
 module.exports = function(passport) {
   passport.use(new FacebookStrategy({
@@ -14,39 +15,27 @@ module.exports = function(passport) {
       clientSecret: process.env.FACEBOOK_APP_SECRET,
       callbackURL:  "http://localhost:3000/auth/facebook/callback"
     },
-    facebookResponse  // see below
+    facebookResponse
   ));
 };
 
-
-function saveResponse(accessToken, refreshToken, profile) {
-  fs.writeFile('../db/oauth_fb_response', {
-    accessToken:  accessToken,
-    refreshToken: refreshToken,
-    profile:      profile
-  });
-}
-
-
 function facebookResponse(accessToken, refreshToken, profile, done) {
-  // DO A BUNCH OF STUFF HERE WITH RETURNED DATA
   User.findOne({facebook_id: profile.id}, function(err, user) {
     if (err) return done('database error');
-    console.log(user);
-    if (user) {
-      getFacebookFriends(user, null, function() {
-        return done(null, user);
-      });
-      return;
+    //If user is already in the database return the user
+    if (user && user.fb_last_update) {
+      return done(null, user);
     }
 
+    //Else create the user
     var newUser = new User();
     newUser.facebook_id = profile.id;
-    console.log(profile.id);
-    // TO BE REMOVED & LOCATED ON CLIENT - ENCRYPTED IN COOKIE
     newUser.access_token = accessToken;
 
-    getFacebookFriends(newUser, null, function(res) {
+    //Populate cards with user's facebook friends
+    getFacebookFriends(newUser, null, function(err, res) {
+      //If error, continue with saving user
+      if (err) { console.log(err); }
       newUser.save(function(err, user) {
         if (err) {
           console.log(err);
@@ -57,30 +46,3 @@ function facebookResponse(accessToken, refreshToken, profile, done) {
     });
   });
 }
-
-function getFacebookFriends(user, url, callback) {
-  url = url || 'https://graph.facebook.com/me/taggable_friends';
-  request
-   .get(url)
-   .query({ access_token: user.access_token })
-   .query({ fields: 'name,picture.width(500)' })
-   .query({ limit: 5000 })
-   .end(function(err, res) {
-      if (err) return console.log(err);
-      var resData = JSON.parse(res.text);
-      var data = resData.data;
-      for (var i=0; i<data.length; i++) {
-        var newCard = new Card();
-        newCard.personName = data[i].name;
-        newCard.personPic = data[i].picture.data.url;
-        newCard.userId = user.facebook_id;
-        newCard.save(function(err, card) {
-        });
-      }
-      if (resData.paging.next) {
-        return getFacebookFriends(user, resData.paging.next, callback);
-      }
-      return callback(res);
-   });
-}
-
