+++ models/User.js
@@ -3,10 +3,10 @@
 var mongoose = require('mongoose');
 
 var UserSchema = mongoose.Schema({
-  eat:          { type: Number                              },
-  access_token: {type: String                               },
-  facebook_id:  { type: String, required: true, unique: true},
-  friends:      { type: Array                               }
+  eat:            { type: Number                               },
+  access_token:   { type: String                               },
+  facebook_id:    { type: String, required: true, unique: true },
+  fb_last_update: { type: Date,                                }
 });
 
 module.exports = mongoose.model('User', UserSchema);
